backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
itsdangerous
slowapi
google-generativeai
